Move cache test
